// src/api/model/sentFlagInvitationResult.ts
/**
 * Generated by orval v8.12.1 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */

export interface SentFlagInvitationResult {
  id?: number;
  flagId?: number;
  flagTitle?: string;
  flagDescription?: string;
  inviteeNickname?: string;
  createdAt?: string;
}